+++ dotnet/samples/Concepts/AgentSyntax/BaseTest.cs
@@ -1,5 +1,4 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
-
 using System.Reflection;
 using Configuration;
 using Microsoft.Extensions.Configuration;
