+++ dotnet/samples/Concepts/AgentSyntax/OpenAIAssistant_CodeInterpreter.cs
@@ -1,30 +1,18 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 using System.Threading.Tasks;
-using Examples;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Agents;
 using Microsoft.SemanticKernel.Agents.OpenAI;
 using Microsoft.SemanticKernel.ChatCompletion;
-using Plugins;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace OpenAIAssistant;
-
+namespace Examples;
 /// <summary>
-/// Demonstrate creation of <see cref="OpenAIAssistantAgent"/> and
-/// eliciting its response to three explicit user messages.
+/// Demonstrate using code-interpreter on <see cref="OpenAIAssistantAgent"/> .
 /// </summary>
-/// <remarks>
-/// This example demonstrates that outside of initialization (and cleanup), using
-/// <see cref="OpenAIAssistantAgent"/> is no different from <see cref="ChatCompletionAgent"/>
-/// even with with a <see cref="KernelPlugin"/>.
-/// </remarks>
-public class OpenAIAssistant_Agent(ITestOutputHelper output) : BaseTest(output)
+public class OpenAIAssistant_CodeInterpreter(ITestOutputHelper output) : BaseTest(output)
 {
-    private const string HostName = "Host";
-    private const string HostInstructions = "Answer questions about the menu.";
-
     [Fact]
     public async Task RunAsync()
     {
@@ -35,14 +23,9 @@ public class OpenAIAssistant_Agent(ITestOutputHelper output) : BaseTest(output)
                 config: new(this.ApiKey, this.Endpoint),
                 new()
                 {
-                    Instructions = HostInstructions,
-                    Name = HostName,
+                    EnableCodeInterpreter = true, // Enable code-interpreter
                     ModelId = this.Model,
                 });
-
-        // Initialize plugin and add to the agent's Kernel (same as direct Kernel usage).
-        KernelPlugin plugin = KernelPluginFactory.CreateFromType<MenuPlugin>();
-        agent.Kernel.Plugins.Add(plugin);
 
         // Create a chat for agent interaction.
         var chat = new AgentGroupChat();
@@ -50,10 +33,8 @@ public class OpenAIAssistant_Agent(ITestOutputHelper output) : BaseTest(output)
         // Respond to user input
         try
         {
-            await InvokeAgentAsync("Hello");
-            await InvokeAgentAsync("What is the special soup?");
-            await InvokeAgentAsync("What is the special drink?");
-            await InvokeAgentAsync("Thank you");
+            await InvokeAgentAsync("What is the solution to `3x + 2 = 14`?");
+            await InvokeAgentAsync("What is the fibinacci sequence until 101?");
         }
         finally
         {
