+++ dotnet/samples/Concepts/AgentSyntax/OpenAIAssistant_Retrieval.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 using System.Threading.Tasks;
 using Configuration;
-using Examples;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Agents;
 using Microsoft.SemanticKernel.Agents.OpenAI;
@@ -11,8 +10,7 @@ using Resources;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace OpenAIAssistant;
-
+namespace Examples;
 /// <summary>
 /// Demonstrate using retrieval on <see cref="OpenAIAssistantAgent"/> .
 /// </summary>
